remove node endpoints from peer lists
