+++ src/Stratis.Bitcoin/Connection/ConnectionManager.cs
@@ -21,67 +21,6 @@ using Stratis.Bitcoin.Utilities.Extensions;
 
 namespace Stratis.Bitcoin.Connection
 {
-    public interface IConnectionManager : IDisposable
-    {
-        /// <summary>
-        /// Adds a peer to the address manager's collection as well as
-        /// the connection manager's add node collection.
-        /// </summary>
-        void AddNodeAddress(IPEndPoint ipEndpoint);
-
-        /// <summary>
-        /// Adds a peer to the address manager's connected nodes collection.
-        /// <para>
-        /// This list is inspected by the peer connectors to determine if the peer
-        /// isn't already connected.
-        /// </para>
-        /// </summary>
-        void AddConnectedPeer(INetworkPeer peer);
-
-        void AddDiscoveredNodesRequirement(NetworkPeerServices services);
-
-        Task<INetworkPeer> ConnectAsync(IPEndPoint ipEndpoint);
-
-        IReadOnlyNetworkPeerCollection ConnectedPeers { get; }
-
-        INetworkPeer FindLocalNode();
-
-        INetworkPeer FindNodeByEndpoint(IPEndPoint ipEndpoint);
-
-        INetworkPeer FindNodeByIp(IPAddress ipAddress);
-
-        INetworkPeer FindNodeById(int peerId);
-
-        string GetNodeStats();
-
-        string GetStats();
-
-        /// <summary>Initializes and starts each peer connection as well as peer discovery.</summary>
-        void Initialize();
-
-        /// <summary>The network the node is running on.</summary>
-        Network Network { get; }
-
-        /// <summary>Factory for creating P2P network peers.</summary>
-        INetworkPeerFactory NetworkPeerFactory { get; }
-
-        /// <summary>User defined node settings.</summary>
-        NodeSettings NodeSettings { get; }
-
-        /// <summary>The network peer parameters for the <see cref="IConnectionManager"/>.</summary>
-        NetworkPeerConnectionParameters Parameters { get; }
-
-        /// <summary>Includes the add node, connect and discovery peer connectors.</summary>
-        IEnumerable<IPeerConnector> PeerConnectors { get; }
-
-        /// <summary>Connection settings.</summary>
-        ConnectionManagerSettings ConnectionSettings { get; }
-
-        void RemoveNodeAddress(IPEndPoint ipEndpoint);
-
-        List<NetworkPeerServer> Servers { get; }
-    }
-
     public sealed class ConnectionManager : IConnectionManager
     {
         /// <summary>Provider of time functions.</summary>
@@ -421,7 +360,18 @@ namespace Stratis.Bitcoin.Connection
             this.logger.LogTrace("({0}:'{1}')", nameof(ipEndpoint), ipEndpoint);
 
             INetworkPeer peer = this.connectedPeers.FindByEndpoint(ipEndpoint);
-            peer?.Disconnect("Requested by user");
+
+            if (peer != null)
+            {
+                peer.Disconnect("Requested by user");
+                this.RemoveConnectedPeer(peer, "Requested by user");
+            }
+
+            this.peerAddressManager.RemovePeer(ipEndpoint.MapToIpv6(), IPAddress.Loopback);
+            List<IPEndPoint> matchingAddNodes = this.ConnectionSettings.AddNode.Where(p => p.Match(ipEndpoint)).ToList();
+            matchingAddNodes.ForEach(m => this.ConnectionSettings.AddNode.Remove(m));
+            var addNodeConnector = this.PeerConnectors.FirstOrDefault(pc => pc is PeerConnectorAddNode);
+            if (addNodeConnector != null) addNodeConnector.MaxOutboundConnections--;
 
             this.logger.LogTrace("(-)");
         }
